refactor: move custom titles folder
